gl: allow running code on the GL context thread
This pigbacks on the existing GL work concept by adding a special
RunOnContextThread() interface method that allows running arbitrary
code in the GL context thread. This is useful if you have a library
that has code that needs to be called in that thread. Note support
for external libraries is not in yet but will come soon.